refactor(discord): streamline auth flow and directory structure
Reorganizes Discord integration files for better maintainability:
- Consolidates Discord connection UI components
- Removes redundant callback pages
- Updates environment configuration
- Refines Inngest function setup for Discord operations
