+++ apps/ai-hero/src/app/discord/page.tsx
@@ -1,10 +1,28 @@
-import { notFound, redirect } from 'next/navigation'
-import { env } from '@/env.mjs'
+import * as React from 'react'
+import { headers } from 'next/headers'
+import { getProviders } from '@/server/auth'
+
+import { DiscordConnectButton } from './discord-connect-button'
 
 export default async function Discord() {
-	if (env.NEXT_PUBLIC_DISCORD_INVITE_URL === undefined) {
-		console.error('Discord invite URL is not set')
-		return notFound()
-	}
-	redirect(env.NEXT_PUBLIC_DISCORD_INVITE_URL)
+	await headers()
+
+	const providers = getProviders()
+
+	const discordProvider = providers?.discord
+
+	return (
+		<main data-login-template="">
+			<h1 data-title="">
+				Join {process.env.NEXT_PUBLIC_PARTNER_FIRST_NAME}{' '}
+				{process.env.NEXT_PUBLIC_PARTNER_LAST_NAME}'s AI Hero Discord
+			</h1>
+
+			<div data-providers-container="">
+				{discordProvider ? (
+					<DiscordConnectButton discordProvider={discordProvider} />
+				) : null}
+			</div>
+		</main>
+	)
 }
