+++ apps/ai-hero/src/env.mjs
@@ -76,6 +76,7 @@ export const env = createEnv({
 		DISCORD_BOT_TOKEN: z.string().optional(),
 		DISCORD_GUILD_ID: z.string().optional(),
 		DISCORD_MEMBER_ROLE_ID: z.string().optional(),
+		DISCORD_SUBSCRIBER_ROLE_ID: z.string().optional(),
 		DISCORD_PURCHASER_ROLE_ID: z.string().optional(),
 		DISCORD_CLIENT_ID: z.string().optional(),
 		CLOUDINARY_API_KEY: z.string(),
@@ -163,6 +164,7 @@ export const env = createEnv({
 		DISCORD_BOT_TOKEN: process.env.DISCORD_BOT_TOKEN,
 		DISCORD_GUILD_ID: process.env.DISCORD_GUILD_ID,
 		DISCORD_MEMBER_ROLE_ID: process.env.DISCORD_MEMBER_ROLE_ID,
+		DISCORD_SUBSCRIBER_ROLE_ID: process.env.DISCORD_SUBSCRIBER_ROLE_ID,
 		DISCORD_PURCHASER_ROLE_ID: process.env.DISCORD_PURCHASER_ROLE_ID,
 		CLOUDINARY_API_KEY: process.env.CLOUDINARY_API_KEY,
 		CLOUDINARY_API_SECRET: process.env.CLOUDINARY_API_SECRET,
